refactored Word and Sentence objects
